enabled api module
=== schema version 2 ===

network.timeout is moved to timeout

network port must *not* be in quotes so it is detected as an interger

incremented client version to 4

default protocol type is now http for development purposes.

network.address is now localhost for development purposes

network.port is 8888 for development purposes, will be a unique port for production.